agents/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# LangGraph + LangChain
langgraph==0.1.5
langchain==0.2.1
langchain-groq==0.1.3
langchain-community==0.2.1
langsmith==0.1.63

# Browser
playwright==1.44.0

# Memory
chromadb==0.5.0
sentence-transformers==3.0.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
redis[asyncio]==5.0.4

# APIs
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.130.0
notion-client==2.2.1
slack-sdk==3.27.1

# Utils
httpx==0.27.0
structlog==24.1.0
python-dotenv==1.0.1
python-telegram-bot==21.2
